pipeline/requirements.txt
openai
python-dotenv
anthropic
supabase>=2.0.0
